# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/comment/createComment": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论业务接口"
                ],
                "summary": "发送评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "发送参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.CreateCommentFrom"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.H"
                        }
                    }
                }
            }
        },
        "/comment/deleteNoteComment": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论业务接口"
                ],
                "summary": "删除评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "发送参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.DeleteNoteCommentFrom"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.H"
                        }
                    }
                }
            }
        },
        "/comment/getCommentList": {
            "get": {
                "tags": [
                    "评论业务接口"
                ],
                "summary": "获取评论记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.H"
                        }
                    }
                }
            }
        },
        "/comment/getNoteCommentList": {
            "get": {
                "tags": [
                    "评论业务接口"
                ],
                "summary": "获取帖子评论列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "node_identity",
                        "name": "note_identity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "utils.CreateCommentFrom": {
            "type": "object",
            "required": [
                "content",
                "note_identity"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "note_identity": {
                    "type": "string"
                }
            }
        },
...
        "utils.DeleteNoteCommentFrom": {
            "type": "object",
            "required": [
                "comment_identity",
                "note_identity"
            ],
            "properties": {
                "comment_identity": {
                    "type": "string"
                },
                "note_identity": {
                    "type": "string"
                }
            }
        },
...
        "utils.H": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}